requirements.txt
backports.ssl-match-hostname==3.5.0.1
requests==2.12.3
six==1.10.0
socketIO-client-2==0.7.2
websocket-client==0.39.0
